Update(PR comments): package rename, refactor of constructors, adding extra variables.
