+++ app/src/main/java/com/m/helper/configModel/config/KeycloakConfig.java
@@ -1,4 +1,4 @@
-package com.m.helper.DTO.config;
+package com.m.helper.configModel.config;
 
 import com.google.gson.annotations.SerializedName;
