+++ app/src/main/java/com/m/helper/configModel/config/SyncConfig.java
@@ -1,4 +1,4 @@
-package com.m.helper.DTO.config;
+package com.m.helper.configModel.config;
 
 public class SyncConfig {
     private String websocketUrl;
